// src/interfaces/dashboard/malmatta-ferfar/FerfarForm.types.ts
export interface FerfarFormData {
  ferfarNaumaYad: string;
  gramPanchayat: string;
  year: string;
  toYear: string;
  anuKramank: string;
  malmattaKramank: string;
  wardNo: string;
  plotNo: string;
  khasaraKramank: string;
  surveyKramank: string;
  masikSabhaKramank: string;
  tharavKramank: string;
  dinak: string;
  gharmalkachiNavLihunDenar: string;
  navLihunGhenara: string;
  sachiv: string;
  sarpanch: string;
  upsarpanch: string;
  sheraTip: string;
}